refactor(ownership): relocate ui to presentation/ui
Matches Car's layering, where interface/ holds schemas only and
presentation/ holds components. Pure relocation plus the import-path
fixes it forces (including the two external overview.tsx consumers);
no behavior changes.
