+++ src/module/car/ownership/presentation/ui/modals/delete/delete.tsx
