+++ src/module/car/ownership/presentation/ui/sections/total-ownerships/total-ownerships.tsx
@@ -2,7 +2,7 @@
 
 import { twMerge } from 'tailwind-merge';
 
-import { useTotalOwnershipsSection } from '@/car/ownership/ui/sections/total-ownerships/use-total-ownerships';
+import { useTotalOwnershipsSection } from '@/car/ownership/presentation/ui/sections/total-ownerships/use-total-ownerships';
 import { DashboardSection } from '@/dashboard/ui/section/section';
 import { Spinner } from '@/ui/decorative/spinner/spinner';
