+++ src/module/dashboard/ui/sections/overview/overview.tsx
@@ -1,4 +1,4 @@
-import { TotalOwnershipsSection } from '@/car/ownership/ui/sections/total-ownerships/total-ownerships';
+import { TotalOwnershipsSection } from '@/car/ownership/presentation/ui/sections/total-ownerships/total-ownerships';
 import { InsuranceExpirationSection } from '@/car/presentation/ui/sections/insurance-expiration/insurance-expiration';
 import { TechnicalInspectionExpirationSection } from '@/car/presentation/ui/sections/technical-inspection-expiration/technical-inspection-expiration';
 import { CostsSection } from '@/car/service-log/ui/sections/costs/costs';
